更新TaiPower模型和DTO中的屬性類型，將decimal類型改為double，以提高數據精度

+++ Dtos/TaiPower/PowerData.cs
@@ -2,15 +2,15 @@ namespace NET_API.Dtos.TaiPower;
 
 public class PowerData
 {
-    public DateTime Time {get; set;}
-    public decimal CentralConsumption { get; set;}
-    public decimal NorthConsumption { get; set;}
-    public decimal SouthConsumption {get; set;}
-    public decimal EastConsumption {get; set;}
+  public DateTime Time { get; set; }
+  public double CentralConsumption { get; set; }
+  public double NorthConsumption { get; set; }
+  public double SouthConsumption { get; set; }
+  public double EastConsumption { get; set; }
 }
 
 public class PowerDataResponse
 {
-  public List<PowerData> Data {get; set;} = new List<PowerData>();
-  public int Count {get; set;} = 0;
+  public List<PowerData> Data { get; set; } = new List<PowerData>();
+  public int Count { get; set; } = 0;
 }
